Added GLSLGenerator.runReferenceImageTest, which computes the reference image for a GLSL feature test in JavaScript. For the special case rendering scenario of the GLSL tests, the results appear to be sufficiently close to those produced on the GPU that the tolerances generally will not need to be adjusted.
Care needed to be taken to evaluate the test function at the correct
locations, and to interpolate the results correctly. Once this was done for
testing of both vertex and fragment shaders, nearly all of the differences
between the CPU and GPU generated results vanished.

Added WebGLTestUtils.setupUnitQuadWithTexCoords and
setupTexturedQuadWithTexCoords in order to be able to conveniently render
the reference images for the vertex shader variants of the tests.

So far, only glsl-function-cos.html has been converted to the new reference
image test runner. After this initial commit, the tests for the other
trigonometric functions will be converted.


git-svn-id: https://cvs.khronos.org/svn/repos/registry/trunk/public/webgl@16909 e7fa87d3-cd2b-0410-9028-fcbf551c1848