Update framebuffer-object-attachment to check for behavior on delete.
The issue is some impls emulate DEPTH_STENCIL_ATTACHMENT
by attaching to both DEPTH_ATTACHMENT and STENCIL_ATTACHMENT
separately. 

You can attach all 3 which would be incomplete but then you
can unattach conflicting ones and it should still work.

If one attachment is unbound they fix internally rebind 
the remaining attachments but if they are instead deleted
they are not rebound. This is a bug.

Updated this test to check for that case.




git-svn-id: https://cvs.khronos.org/svn/repos/registry/trunk/public/webgl@18137 e7fa87d3-cd2b-0410-9028-fcbf551c1848