Add test for using too many uniforms.
I'm not 100% sure this test is valid. The test could fail
because the shader is too complicated. Though it seems like
it's a simple example of using all uniforms and if you can't
compile a shader using all advertised uniforms then maybe
that implemention shouldn't be advertising that many uniforms?