Speed up the quickCheckAPI tests
The issue is that for some reason Chrome is EXTREMELY slow
at laying out unprintable characters. While there's no excuse
that it's slow there's also not much insentive to fix it
since few pages use unprintable characters. So, sanitize the
output so no unprintible characters are added to the DOM.