Fixed problems in reference image generation for GLSL vertex shader tests.
glsl-function-smoothstep-float should use high precision for OpenGL ES
https://www.khronos.org/bugzilla/show_bug.cgi?id=728

For those GLSL function tests which have been migrated to generate reference images in JavaScript, changed the vertex shader variant to use the same geometry for the reference and test shaders. The previous approach of generating the reference image into a texture was incorrect because the two versions used very different interpolation techniques. Thanks to Olli Etuaho from NVIDIA for diagnosing the problem and Chris Green from Imagination for initially raising the issue.

This is a bug fix to the conformance suite, so the fix has been backported as far back as version 1.0.1.