Fix interpolation in reference image generation for GLSL fragment shader tests
Emulate the barycentric coordinate interpolation done by the GPU in the
reference renderings for the fragment shader tests. Previously color
values for fragments were computed per-fragment using the bilinearly
interpolated texCoord value, which did not match up with the barycentric
interpolation from the 3 nearest vertices' color values.

This is a bug fix to the conformance suite, so the fix has been
backported as far back as version 1.0.1.