Fix interpolation problems in GLSL fragment shader tests
In the test rendering, the color varying value was being interpolated by
the GPU within each triangle by using barycentric coordinates (as per
GLES2 spec section 3.5.1) prior to being operated on in the test fragment
shaders. This did not match up with the reference rendering, where this
color value is calculated per fragment using the bilinearly interpolated
"texcoord" value.

Make the reference and test shaders agree by calculating the color value
per-fragment also in the test shaders. This can be done by setting the
texcoord into a varying. Since the texcoord is simply a multiple of the
vertex position, interpolating it by using a varying agrees with bilinear
interpolation.

This is a bug fix to the conformance suite, so the fix has been
backported as far back as version 1.0.1.