Enforce MAX_TEXTURE_SIZE to be a power of two
This is implied by GLES2.0 spec section 3.7.1. There, the expression
"level zero through k" strongly suggests that k is expected to be an
integer, and it follows that MAX_TEXTURE_SIZE should be a power of two.
This holds in practice for almost all current implementations. Clarify
this interpretation of the spec by explicitly testing it.