Clean up webgl-test-utils.js
Fix behavior when automatically determining the type of a shader.
Previously the second shader in setupProgram could get assigned the type
of the first shader in case only the type of the second shader was
specified in its script tag, possibly resulting in confusing errors.

The shaderType logic in loadShaderFromScript was also needlessly
complicated, so simplify it.

In addition, fix some typos, add missing docs, and remove duplicate of
setupSimpleTextureProgram.

These fixes are meant to ease development of new tests, and aren't ported
back to earlier versions of the test suite.