Fix antialias setting in copy-tex-image-and-sub-image-2d.html
Previously, if the antialias flag was set, the test actually forced
antialiasing off. If antialias flag was not set, the test used the
default setting in WebGL test utils, which also is antialiasing off. Fix
this so that antialiasing is always explicitly specified in the test and
that having antialiasing on gets tested. This fix is backported to 1.0.1
and 1.0.2.

Also fix formatting of this test's output in 1.0.3+