Clean up fullscreen related functions in webgl-test-utils.js
Remove unnecessarily complicated findElement sub-function, make
fullscreenStateNames private to the function that uses it, put docs where
they belong and add in missing docs.

Only manual/framebuffers-keep-contents-exiting-fullscreen-mode.html is
using this functionality of the test utils, and has been tested to work
with these changes.

This change is meant to ease development of new tests, and is not ported
back to earlier versions of the test suite.