Add a checkbox for auto-scrolling in the test runner
Sometimes, it can be useful to be able to inspect previous test results
while tests are still running. Add the option to disable auto-scrolling
the test list in the test runner for this purpose.

The patch is made to 1.0.2 and later versions of the suite, where the
test runner is identical.