Limit memory usage in webgl-compressed-texture-size-limit.html test
This case is more complicated than texture-size-limit, since null can not
be passed to compressedTexImage2D, and so an ArrayBuffer with the correct
size needs to be created to test the size limits. And the texture image
must be square in order to properly test the cube map. Either way, this
patch cuts memory usage of the largest buffer nearly by a factor of four,
which can turn a 1 GB allocation into a more manageable 256 MB allocation
on hardware supporting max texture width of 16384 pixels, for example.

Still, even with this change this test can be a source of failures
related to memory allocation rather than texture size limits.