Fix to adhere to the row/column confusion in J3DIMath
In J3DIMatrix, $matrix.mxy is actually the matrix value in the xth column
and yth row. WebKitCSSMatrix, which J3DIMatrix is based on, has incorrect
documentation about this. J3DIMatrix documentation in about row/column
majorness might be wrong as well, but this patch only attempts to fix the
matrix decomposition.