Fix remaining incorrect math in rotation decomposition
The cos(rotate[1]) != 0 code path in rotation decomposition used to work
only for a special case of the rotation. Fix it so that it should work
with arbitrary rotation.

The fix is demonstrated in MatrixTest.html. The first cube and the cube
with the decomposed rotation should rotate roughly the same way.