Fix the most obvious issues in J3DIMath and add basic tests
Interface description was edited to match what the code does, cross
product now produces the correct results, and basic unit tests for
the functions where problems were found were added.

Note that WebKitCSSMatrix behavior does not seem to match docs in current
browsers: m12 represents the value in the 2nd row, not in 2nd column like
the class reference says [1]. The actual behavior is similar to how
J3DIMath implements it in JavaScript.

[1] See https://developer.apple.com/library/safari/documentation/AudioVideo/Reference/WebKitCSSMatrixClassReference