Add proposal for WEBGL_draw_elements_no_range_check
The goal of the extension is to eliminate one common source of WebGL CPU
overhead by removing drawElements index validation.

The extension changes the behavior of drawElements as opposed to
introducing a new call or a toggleable flag to make it possible to get
rid of index buffer copies in CPU memory entirely after the extension
has been enabled.