Add simple benchmark for drawElements index validation overhead
The test changes data in an index buffer and calls drawElements on it.
There are always some invalid indices in the buffer, so that any drawing
is not performed but the test just measures the overhead from uploading
data and performing validation.

The test is written to specifically target implementations in Chrome and
Firefox in March 2014, so some of the results might be slightly
misleading on other implementations.