Add out-of-bounds tests for ANGLE_instanced_arrays extension
This is achieved by refactoring existing draw out of bounds tests to
resources/out-of-bounds-test.js so that they can be used both by regular
draw functions and their instanced variants.

Additionally, move tests that don't have anything to do with
out-of-bounds checks from draw-arrays-out-of-bounds.html and
draw-elements-out-of-bounds.html to gl-drawarrays.html and
gl-drawelements.html.

Some tests are added to unify testing done to drawArrays and
drawElements.