Do not show 1.0.1 and 1.0.2 test subsets in version selector
The version selector in the latest version of the conformance suite seems
to be misleading. It suggests to people new to the conformance suite that
they're running the 1.0.1 or 1.0.2 tests, when in fact some of the tests
in the subset might be updated to increase coverage after the old version
of the test suite was frozen. Remove 1.0.1 and 1.0.2 from the selector to
reduce confusion.