Add extra ANGLE test for scoped structures.
Because ANGLE internally uses a renaming scheme for scoped structures,
it's possible for the user to make certain structurse that conflict
with the internal scheme and cause link errors.

This bug only manifests in Firefox, not Chrome, because Chrome uses a
preliminary shader rewriting pass that hashes all variable names. It is
still possible to trigger from Chrome internal shaders and with
ANGLE as a standalone driver.