Added errorbound to check on pixel center
Added errorbound to check on pixel center.
Do you think it is needed ? I thought about it but decided not to add
the error bound because 0.5 is exact representable in binary (0.1). Even
if there is a gpu which does use an internal representation which can
not exactly represent 0.5 (dec) it should compute a representation which
is the closest it can get for both the fractional coordinate and for the
compiled literal. So the comparison should still succeed. Is there a
central place where I can post these kind of technical questions to
vendors ?