Remove DrawRangeElements from WebGL 2
WebGL would require that the implementation handles indices that are
outside the start, end range in a consistent way, but the OpenGL ES 3.0
specification explicitly states that behavior in this case is
implementation-dependent. Because of this, WebGL would need to validate
that the indices are in the start, end range to ensure consistent
behavior, and no performance benefits could be realized by exposing this
function.