Fix typos in extension tests and add #extension directive test
Testing the #extension directive separately makes it more obvious where a
bug in a WEBGL_draw_buffers implementation might be. Firefox currently
seems to be suffering from a bug where WEBGL_draw_buffers is exposed on
all GLES3-capable devices without checking that they support the GLES2
extension. On a careful reading of the GLES3 spec section 4.2.1, it seems
that multiple draw buffers are not supported for GLSL ES 1.0 shaders on
vanilla GLES3, even though multiple draw buffers are otherwise part of
the specification.