Specify an error for setting line width to NaN
OpenGL drivers are allowed to generate errors when NaN is supplied to
functions under GLES 3.1 section 2.3.3.1 or OpenGL 4.5 section 2.3.4.1.
Similar clauses exist in earlier versions of the specs as well. NVIDIA
OpenGL driver generates an error when NaN is supplied to glLineWidth.
Make this behavior consistent for WebGL, which removes the possibility of
drawing with an undefined line width.

Functions which supply data to shaders are not changed, as this would
have an undesirable performance cost and only a small benefit, since it
would not prevent shaders themselves to generate NaNs.