Test that setting line width to NaN is not an error
This exposes an NVIDIA driver bug, where INVALID_VALUE is incorrectly
generated. glLineWidth should only produce an error when width <= 0,
which is false for width NaN.

Rendering with width NaN is undefined, so alternatively the spec could be
changed to disallow setting it, but this is unlikely to become a major
issue.