Partially revert "Test that setting line width to NaN is not an error"
This turned out to be not a driver bug after all, but within allowed
behavior. GL specs say that providing NaN or infinity to a function
accepting floating point data yields unspecified results (GLES3.0 2.1.1).
So actually, generating an error is allowed in this case.

Keep the part of the commit that fixed a comment in the test file.

This reverts commit cf16a1cae5179c76397e7bb6f56e6e91dd0cf650.