Remove samplerParameter functions taking multiple values
There is no foreseeable need for these functions. They're not used in any
existing ES extensions, and even though texture swizzling is in the ES3.0
spec, it only supports setting the swizzle values one by one, and not
with a single call using TEXTURE_SWIZZLE_RGBA.

This is consistent with the removal of similar texParameter functions in
WebGL 1.0.