Change DEPTH_BITS for UNSIGNED_INT from 24 to 16.
The depth texture extension says the following with regard to UNSIGNED_INT depth textures.
"As per the OpenGL ES spec, there is no guarantee that the OpenGL ES implementation will use the to determine how to store the depth texture internally. It may choose to downsample the 32-bit depth values to 16-bit or even 24-bit."

For the sake of not imposing new requirements, we'll change the minimum depth for UNSIGNED_INT from 24 to 16.