Add information on planned backwards incompatibility in WebGL 2
It was decided in the WebGL WG F2F that WebGL 2 will not be entirely
backwards compatible with WebGL 1. In particular, there is agreement
that all extensions are not moving over, and always enabling primitive
restart in WebGL 2 significantly streamlines implementations, so that
could be added to this section of the spec.