Modify test runner to dump translated shaders
This change adds a test runner feature that dumps translated shaders
while running glsl conformance tests.

For each test, a summary json file is created which has all the
information like original shaders, test description, whether shader
should compile and path to translated shaders.

For all the glsl test shaders that compile successfully, translated
shaders are dumped to separate files so that it can be fed directly
to the compiler.

All the translated shader and summary files are compressed into a
tar.gz at the time of finishing tests.

New test runner command line option --dump_shaders is added and it
needs to be passed for dumping translated shaders.