bindAttribLocation with aliasing
This test verifies combinations of valid, active attribute types cannot be bound to
the same location with bindAttribLocation.

Per the OpenGL spec: It is possible for an application to bind more than one attribute
name to the same location. This is referred to as aliasing. This will only work if
only one of the aliased attributes is active in the executable program, or if no path
through the shader consumes more than one attribute of a set of attributes aliased to
the same location.

vertex shaders take the form:
precision mediump float;
attribute $(type_1) a_1;
attribute $(type_2) a_2;
void main() {
    gl_Position = $(gl_Position_1) + $(gl_Position_2);
}