Add notes on backwards incompatibility in error cases in WebGL 2
As in GLES, behavior in the cases where an earlier API version generates
an error is subject to change in future API versions. Also, in some cases
GLES3.0 can sample a texture even though GLES2.0 would return RGBA =
0, 0, 0, 1.

It seems like a good idea to mention this kind of things in the backwards
incompatibility section for the sake of exhaustiveness, though a complete
list of changed error conditions would seem excessive.