Improve the explanation of viewport and scissor in the WebGL 1 spec
The spec mentioned that drawing was constrained by canvas area, but it
should instead refer to framebuffer area, since rendering to an FBO
should not be constrained by the size of the canvas. Also, clarify that
the viewport only constrains the drawing of primitives, and not clearing
the framebuffer. This should remove the most glaring contradictions with
the GLES2.0 spec.

A mention of how the drawing buffer size is initialized is removed from
this section, as it was redundant and confusing.

Add a clear mention that drawing commands can only modify pixels inside
the currently bound framebuffer, since the GLES2.0 spec does not seem to
clearly address what happens when the viewport extends outside the
framebuffer and draw commands are called.