Modify xsl etc. to allow extensions to add new interfaces. These can be entered in the <newfun> sections. Use these in WEBGL_dynamic_texture in place of comment that such functionality is needed.
The CSS for dt.idl-code in default.css has been removed. This exdented the first line of text by an amount equal to the margin it was setting. The exdent made the multi-line interface definitions look ugly. Removing this means all lines begin at the same offset where the first line would have been. All other uses of <dt class="idl-code"> in this repo are single line so are not affected by this change.

This commit merges https://github.com/msc-hi/WebGL/pull/1.