Add a test for constant variable precision qualifiers
The test checks that a precision qualifier of a constant variable affects
the precision of a consuming operation. It targets platforms which
implement different precisions for mediump and highp in hardware, it
will pass on all hardware that only has one precision corresponding to
highp. Also if the test is run on a platform that does not support highp,
it passes without running the shader.