Remove DataView Constructor
Since the WebGL spec was first authored, the TypedArray spec has been incorporated
into the ES6 draft spec linked at:
http://wiki.ecmascript.org/doku.php?id=harmony:specification_drafts.

data-view-test claims that DataView used as a function should throw an exception. However,
according to the newest draft of ES6, DataView is allowed to be called as a function
as well as a constructor. See section "24.2.2 The DataView Constructor"

Since no WebGL functionality is directly used this particular test, we should remove it
from the WebGL conformance suite and handle it as part of the ECMAScript Conformance Test
Suite.