Loosen fragment shader tolerance in glsl function tests from 0 to 1
Running the tests with precision emulation enabled revealed slight
differences in results from built-in functions compared to their emulated
counterparts. It is possible that these differences reproduce also on
spec-compliant hardware if a built-in function is compiled to different
assembly instructions with higher internal precision than its emulated
counterpart, for example. Because of this, tolerances for GLSL function
tests need to be loosened for calculations done in mediump.

This change is backported to previous versions of the conformance suite.