Reduce memory usage of multisample corruption test
Multisample corruption tests create a big canvas and at every
iteration test has to allocate a very big buffer for checking drawn
rectangle color. With large number of iterations, if these buffers
are not garbage collected in time, browser memory utilization goes
very high and browser even hangs or gets killed on platforms like
android.

This change modifies tests to allocate a read back buffer in advance
and reuse it for all iterations.

The fix is backported to 1.0.2 and 1.0.3 test suites.