Angle-instanced-arrays calls notifyFinishedToHarness multiple times
Conformance tests notify the harness of completeness one of two ways:
1) Manually include "var successfullyParsed = true;" and js-test-post.js at
   the end of the test. js-test-post.js calls notifyFinishedToHarness
2) Call finishTest at the of the test. FinishTest performs 1 programatically.

Angle-instanced-arrays does both of these things. The "TEST COMPLETED" message shows
up twice in the final output.

Since finishTest is called as part of multiple setTimeout callbacks, we end up
notifying the harness of completeness (through the manually inserted script tag)
before the test actually finishes.

The fix is to delete js-test-post.js from the end of the test and rely on finishTest
exclusively.