Buffer calls to console.log to reduce amount of debugging output.
Only output the buffered logs if the test fails -- i.e., calls testFailed().

This reduces the amount of debugging output for a typical run of the
conformance suite, while still remaining debuggable in case of errors.

Tested by injecting a synthetic failure in a test which previously had
debugging output; verified that the output was printed, and only then.