Test runner: Support Chrome on Android
This patch adds support for running tests remotely on a connected Android
device. It adds support for Chrome on Android and the framework can be
reused for other browsers on Android.

For this, the test runner is refactored so that browsers running locally
and browsers running on an Android device are abstracted behind a
BrowserInstance interface.

To run the tests on Android, you need the following:
    - Android device connected to linux host machine
    - Chrome application installed on the Android device
    - Android SDK installed on your host linux machine
    - 'adb' in your system path
    - The 'ant' build system installed in case you want to modify the
      prebuilt port forwarder apk.

Android support uses adb port forwarding from the connected device
to the host linux machine. When running the tests on Android, the test
runner starts a port forwarder service on the device, and then launches
Chrome on Android using adb to start the tests.

As the port is being forwarded from device to host, all browser requests
to localhost:<port> on the Android device will be forwarded to
localhost:<port> on host linux machine where the local server is set up
by test runner to host the tests.

Along with the source, prebuilt port forwarder service apk is also added
to the tree.

The platform argument needs to be explicitly passed as 'remote-android'
to run tests remotely on android.
Example:
  node ./test-runner --browser Chrome --platform remote-android

Once the tests finish running, results are posted back to the host linux
machine, port forwarder service and browser are terminated, and other
necessary cleanup is done.

This commit also fixes a bug where a wrong directory name was used for
dumped shaders when the browser was not specified on the command line.

This patch was developed in collaboration with Arun Patole
(apatole@nvidia.com).