Fix global initializer containing an uniform in conformance tests
ESSL 1.00 section 4.3 says that global initializers need to be constant
expressions. It is still being discussed how strictly this rule should
apply to WebGL, but initializing globals with uniforms is clearly not
allowed by the spec, so fix the conformance test doing that.

The fix is backported all the way to 1.0.1 conformance suite.