Loosen global variable initialization spec to meet implementations
Historically WebGL implementations have not posed many restrictions to
operations or variables in global variable initializers. This is a
violation of the OpenGL ES Shading Language spec.

Specify what is allowed in global variable initializers in WebGL.
The new spec is halfway between the actual ESSL spec and the old
de-facto functionality. It is intended to avoid widespread compatibility
regressions in browsers while bringing WebGL closer to the restrictions
in the ESSL spec.