Add different versions of constant-precision-qualifier test
After fixing the bug in ANGLE that made this test fail, it was discovered
that some NVIDIA mobile drivers also have a bug making this test fail.
Add two more versions of the test, one that doesn't use constants and one
that only uses highp variables, so it's easier to narrow down
platform-specific failures of this test.