Make ARRAY_BUFFER enum tests conditional on context version
Some of the enums that were assumed to be invalid desktop enums in the
WebGL 1 test are actually added in WebGL 2. The tests are moved from
invalid-passed-params.html to gl-enum-tests.js to leverage existing
context version checks there.

The calls that were testing valid parameters in
invalid-passed-params.html are removed entirely - this doesn't cost test
coverage, since there are plenty of tests for them elsewhere.