Don't remove non-whitespace characters when stripping indentations
Some dEQP test shaders are indented in an inconsistent way: for example,
some lines might be indented with <tab><tab> whereas others have
<tab><space><tab>. These two indentations may look the same in an editor,
but when the framework indentation removal function looks at the
indentations, it counts <tab> as indentation by 4 and <space> as
indentation by one. The first line is used as reference and the rest of
the lines get the same amount of indentation removed. With inconsistent
indentation in the source, this could remove some meaningful characters,
like turning "void main()" into "oid main()" unless it is separately
checked that no non-whitespace characters are removed.

This fixes several shader tests, particularly in
deqp/data/gles2/shaders/constants.html.