Fix draw referencing a texture that was never bound
This test was accidentally rendering with the 0 texture, which worked
because it was expecting black to be drawn. This exposed a bug in
Chrome, and a separate test has been created to provoke it and ensure
drawing with texture 0 results in black, but for this test it's better
for the draw call to explicitly draw a black quad without unnecessary
texture sampling.