Remove dEQP #extension directive test to keep backwards compatibility
When ANGLE implemented a fix to this test in May 2015, significant issues
with backwards compatibility were detected. It was decided that the ESSL1
compiler should issue a warning rather than an error in this case, though
the WebGL spec was not updated.

The spec is still strictly enforced in the case of ESSL3 shaders and
WebGL 2 with a similar gles3 dEQP test.

The WebGL specs are updated to lock down the behavior for both ESSL1
(backwards-compatible behavior) and ESSL3 (strict interpretation of the
spec).

Another alternative was considered where "disable" directives would be
subject to stricter rules closer to the GLSL ES 1.00 spec, but this did
not seem to be useful. This does leave open the possibility of using
extension directives in a confusing way in ESSL1 shaders, but in practice
this does not seem to be a problem for developers.