A query's result must not become available the same frame it was issued.
This specification language is derived from that in the EXT_disjoint_timer_query extension. It is intended to ensure best portability and performance of code using query objects. The ported dEQP tests in the WebGL 2.0 conformance suite have already been adjusted to follow this requirement. The EXT_disjoint_timer_query conformance test will be updated shortly to follow the same requirement.

Per feedback from @RafaelCintron , added non-normative text describing when the "current frame" advances. It is likely that this will have to be updated again in the future.