Update dEQP shader preprocessor tests from deqp-dev
Pull in latest changes to the tests from deqp-dev branch.

This contains the following changes:

-Double underscores in macro names are now expected to be accepted by the
 compiler. See previous discussion in pull request #1230.
-Unrecognized pragmas are now expected to be accepted by the compiler.
 ESSL 1.00 and 3.00 specs say "If an implementation does not recognize
 the tokens following #pragma, then it will ignore that pragma".
-Non-matching varying invariance is fixed in some #pragma tests.
-Added tests for "defined" operator corner cases for ESSL 3.00 (now
 covered by WebGL 1.0 spec section 6.31)
-Added tests for backslash in a comment for ESSL 3.00

gles2/shaders/preprocessor.test is not a direct copy from dEQP, since
WebGL-specific changes apply to the behavior of extension directives.

dEQP scripts are updated with support for "build_successful" expectation.
Also, testFailed() is now called if the shader file can't be parsed to
get debugging information automatically.